+++ UnitedOperationsFramework/AIDriversModule/cfgFunctions.hpp
@@ -1,9 +1,10 @@
 class UO_FW_AIDrivers {
 	file = "x\UO_FW\addons\main\AiDriversModule\cfgFunctions\common";
-	class aiDriversInit {postInit = 1;};
+	class aiDriversInit {};
 	class aiDriversToggle {};
 	class aiDriversRemoveUnit {};
 	class aiDriversCreateUnit {};
 	class aiDriversToggleDriverCam {};
 	class aiDriversEnableAIDriver {};
+	class aiDriversVehInit {};
 };