+++ UnitedOperationsFramework/HostageModule/cfgFunctions.hpp
@@ -1,6 +1,6 @@
 class UO_FW_Hostage {
 	file = "x\UO_FW\addons\main\HostageModule\cfgFunctions\common";
-	class HostageInit {postInit = 1;};
+	class HostageInit {};
 	class HostageSet {};
 	class HostageRescue {};
 	class HostageIsRescued {};