+++ CHANGELOG.md
@@ -9,8 +9,10 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added Button to setup respawn and settings in 3den
 - Added Map Compass Remover
+- Added Hostage Module
 ### Changed
 - Removed briefing limitations
+- Made more Modules independent
 ### Fixed
 - Ai Driver wasn't being added when ace menued
 ### Removed